backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
httpx==0.27.2
celery==5.4.0
redis==5.1.1
anthropic==0.36.2
Pillow==10.4.0
beautifulsoup4==4.12.3
python-multipart==0.0.9
pydantic-settings==2.5.2
